feat: rate mechanic

+++ backend/internal/handlers/mechanics/types.go
@@ -47,7 +47,7 @@ type MechanicUpdate struct {
 }
 
 type RateMechanicParams struct {
-	Rating int `validate:required bson:"rating" json:"rating"`
+	Rating float64 `validate:required bson:"rating" json:"rating"`
 }
 
 type RequestType string
